#2 Remove tests for number of unique TFE groups
This commit removes failing tests designed to verify that the number of unique t groups in the raw data is the same as the number of unique t groups in the estimated model. R and STATA omit some t groups to prevent collinearity. The number of unique groups in eventstudyr matches the number used in base STATA (base STATA includes an additional group as a constant). Coefficients and standard errors agree between base STATA and eventstudyr, so we feel confident that t groups are being handled correctly