Only import individual functions #28
See https://r-pkgs.org/dependencies-in-practice.html#in-code-below-r for
the motivation behind only importing the used functions instead of the
entire package. I left dplyr, ggplot2, and stringr alone because we use
a number of functions in different scripts. I left estimatr because
I was getting a warning with `@importFrom estimatr tidy` since estimatr
uses tidy but doesn't actually create it, see here
https://github.com/DeclareDesign/estimatr/blob/1a0131a2cb8a891cb902c875c680e8f2228a6651/R/S3_tidy.R